Refactor with StreamValueObject base
