+++ source/OutgoingMessages.Domain/OutgoingMessages/Queueing/MarketDocumentStream.cs
@@ -12,41 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using System.IO;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.MarketDocuments;
 
 namespace Energinet.DataHub.EDI.OutgoingMessages.Domain.OutgoingMessages.Queueing;
 
 #pragma warning disable CA1711 // Is a "Stream" value object
-public record MarketDocumentStream : IMarketDocumentStream
+public record MarketDocumentStream : StreamValueObject, IMarketDocumentStream
 {
-    private readonly Stream _stream;
-
     public MarketDocumentStream(IArchivedFile archivedFile)
-        : this(archivedFile?.Document.Stream)
+        : base(archivedFile?.Document.Stream)
     { }
 
     public MarketDocumentStream(FileStorageFile fileStorageFile)
-        : this(fileStorageFile?.Stream) { }
+        : base(fileStorageFile?.Stream) { }
 
     public MarketDocumentStream(MarketDocumentWriterMemoryStream stream)
-        : this((Stream)stream)
+        : base(stream)
     { }
-
-    private MarketDocumentStream(Stream? stream)
-    {
-        ArgumentNullException.ThrowIfNull(stream);
-        _stream = stream;
-    }
-
-    public Stream Stream
-    {
-        get
-        {
-            _stream.Position = 0;
-            return _stream;
-        }
-    }
 }
